// src/enum/user.enum.ts
export enum EUserStatus {
	ACTIVE = 'ACTIVE',
	BLOCKED = 'BLOCKED',
	DISABLED = 'DISABLED',
}
export enum ESearchFields {
    DisplayName = 'displayName',
    Email = 'email',
    Phone = 'phone',
    IdentifyingNumber = 'identifyingNumber',
}
export enum EUserSubject {
	INDIVIDUAL = 'INDIVIDUAL',
	LEGAL = 'LEGAL',
}